+++ mobile/android/android-components/components/concept/engine/src/main/java/mozilla/components/concept/engine/EngineView.kt
@@ -2,12 +2,17 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
+@file:Suppress("Deprecation")
+
 package mozilla.components.concept.engine
 
 import android.graphics.Bitmap
 import android.view.View
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.OnLifecycleEvent
+import mozilla.components.concept.engine.EngineView.InputResult.INPUT_RESULT_HANDLED
+import mozilla.components.concept.engine.EngineView.InputResult.INPUT_RESULT_HANDLED_CONTENT
+import mozilla.components.concept.engine.EngineView.InputResult.INPUT_RESULT_UNHANDLED
 import mozilla.components.concept.engine.selection.SelectionActionDelegate
 
 /**
@@ -91,7 +96,13 @@ interface EngineView {
     /**
      * @return [InputResult] indicating how user's last [android.view.MotionEvent] was handled.
      */
+    @Deprecated("Not enough data about how the touch was handled", ReplaceWith("getInputResultDetail()"))
     fun getInputResult(): InputResult = InputResult.INPUT_RESULT_UNHANDLED
+
+    /**
+     * @return [InputResultDetail] indicating how user's last [android.view.MotionEvent] was handled.
+     */
+    fun getInputResultDetail(): InputResultDetail = InputResultDetail()
 
     /**
      * Request a screenshot of the visible portion of the web page currently being rendered.
@@ -134,6 +145,7 @@ interface EngineView {
      * @see [INPUT_RESULT_HANDLED]
      * @see [INPUT_RESULT_HANDLED_CONTENT]
      */
+    @Deprecated("Not enough data about how the touch was handled", ReplaceWith("InputResultDetail"))
     enum class InputResult(val value: Int) {
         /**
          * Last [android.view.MotionEvent] was not handled by neither us nor the webpage.